Implement loading details of all listings at once
Prior to this change, listings from only the first page of the search
results were being loaded.

This change implements the loading logic in such a way that all the
results from all the pages are loaded at once.

However, the solution is not optimum. The first page has to load twice
in order to get the total number of listings.