Remove `@apollo/server`, `copyfiles` from `server`
Since `@apollo/server` does not support Hapi.js yet, it has been removed
from the `server` package. `copyfiles` has also been removed, since the
graphql file is not needed to be copied to the output directory. Hence,
the `build` script for the `server` package is also updated.

However, the commented lines of code in the server program still remain.