build: fix Docker configuration
Due to files being rearranged by 7f5cc82, file and directory paths in
the Dockerfile and the Docker Compose file needed fixes.

Moreover, separation of logic (via Yarn workspaces and Lerna) into the
new `common` package by b907d68 meant that simply copying `server/out/`
and running it with the `package.json` file of the `server` package was
not going to work. Furthermore, avoiding having to install Lerna was
necessary, since it takes up about 92M, which could increase the size of
the Docker image by a whopping 50%!

For this reason, a few extra lines were added in the Dockerfile, which
in effect added a few layers in the image. But the size of the image has
not increased (noticeably), since no new code or package has been added.

Just as before, the build process is multi-stage, which might leave a
few unused, dangling images. These can later be cleaned up by running
the command `docker image prune`.